refactor: Document 테스트 util 패키지 생성
